refactor(app): split src/cli/app/mod.rs into feature submodules
